+++ lib/node/state.d.ts
@@ -3,6 +3,12 @@
 // Definitions by: David Mesquita-Morris <http://state.software>
 // Definitions: https://github.com/DefinitelyTyped/DefinitelyTyped
 
+export interface Behavior {
+    (message: any, instance: IInstance): void;
+}
+export interface Guard {
+    (message: any, instance: IInstance): boolean;
+}
 export declare enum PseudoStateKind {
     Choice = 0,
     DeepHistory = 1,
@@ -16,6 +22,7 @@ export declare enum TransitionKind {
     Local = 2,
 }
 export interface Element {
+    getAncestors(): Array<Element>;
     getRoot(): StateMachine;
     toString(): string;
 }
@@ -25,6 +32,7 @@ export declare abstract class NamedElement<TParent extends Element> implements E
     static namespaceSeparator: string;
     readonly qualifiedName: string;
     protected constructor(name: string, parent: TParent);
+    getAncestors(): Array<Element>;
     getRoot(): StateMachine;
     toString(): string;
     accept<TArg>(visitor: Visitor<TArg>, arg?: TArg): void;
@@ -45,13 +53,23 @@ export declare class Vertex extends NamedElement<Region> {
 export declare class PseudoState extends Vertex {
     readonly kind: PseudoStateKind;
     constructor(name: string, parent: Region | State | StateMachine, kind?: PseudoStateKind);
+    isHistory(): boolean;
+    isInitial(): boolean;
     accept<TArg>(visitor: Visitor<TArg>, arg?: TArg): void;
 }
 export declare class State extends Vertex {
     readonly regions: Region[];
     defaultRegion: Region;
+    entryBehavior: Behavior[];
+    exitBehavior: Behavior[];
     constructor(name: string, parent: Region | State | StateMachine);
     getDefaultRegion(): Region;
+    isFinal(): boolean;
+    isSimple(): boolean;
+    isComposite(): boolean;
+    isOrthogonal(): boolean;
+    exit(action: Behavior): this;
+    enter(action: Behavior): this;
     accept<TArg>(visitor: Visitor<TArg>, arg?: TArg): void;
 }
 export declare class StateMachine implements Element {
@@ -59,8 +77,10 @@ export declare class StateMachine implements Element {
     readonly regions: Region[];
     defaultRegion: Region | undefined;
     clean: boolean;
+    private onInitialise;
     constructor(name: string);
     getDefaultRegion(): Region;
+    getAncestors(): Array<Element>;
     getRoot(): StateMachine;
     accept<TArg>(visitor: Visitor<TArg>, arg?: TArg): void;
     toString(): string;
@@ -69,10 +89,14 @@ export declare class Transition {
     readonly source: Vertex;
     readonly target: Vertex;
     readonly kind: TransitionKind;
-    guard: (message: any, instance: IInstance) => boolean;
+    guard: Guard;
+    effectBehavior: Behavior[];
+    private onTraverse;
     constructor(source: Vertex, target?: Vertex, kind?: TransitionKind);
-    when(guard: (message: any, instance: IInstance) => boolean): this;
+    when(guard: Guard): this;
+    effect(action: Behavior): this;
     accept<TArg>(visitor: Visitor<TArg>, arg?: TArg): void;
+    toString(): string;
 }
 export declare class Visitor<TArg> {
     visitElement(element: Element, arg?: TArg): void;