+++ test/unit_model.js
@@ -1,8 +1,9 @@
-var state = require("../../lib/node/state");
+var assert = require("assert"),
+	state = require("../lib/node/state");
 
 var model = new state.StateMachine("model");
 var initial = new state.PseudoState("initial", model);
-var stateA = new state.State("stateA", model);
+var stateA = new state.State("stateA", model).exit(function () { console.log("Exit A"); });
 var stateB = new state.State("stateB", model);
 
 initial.to(stateA);
@@ -10,11 +11,9 @@ stateA.to(stateB).when(function (message) { return message === "move" });
 
 var visitor = new state.Visitor();
 
-model.accept(visitor);
-
-
-
-
+stateA.getAncestors().forEach(function (element) {
+	console.log(element + " is ancestor of " + stateA);
+});